Minor fix to random cuboid point function.
The Vector3.RandomPointInCuboid function was using too large of values
for the random offsets, so they were often outside the cuboid. This
change ensures that the values are now in the correct range. Also
decoupled the random point functions a bit so they're now relative to
the origin, and it's the caller's job to do the vector add.