Added depth of field, updated Readme.
I finally figured out how to do depth of field in a ray tracer, and it
has turned out very nice. As with most effects in rendering, it
increases the render time by quite a bit, but it's well worth the wait.
I also added an image that showcases the new effect.