Revised depth of field aperture.
The circle used for distributing points over with depth of field was
previously parallel to the camera plane for all ray directions, but this
was causing some small declines in quality closer to the edges of the
screen. These changes now make the circle's normal parallel to the ray,
and depth of field should appear more correct now.