Added PostProcessing type, other syntax changes.
The PostProcessing module holds functions for a few different pixel
operations. It's just used for clamping the output image right now, but
it also has some rough draft implementations of bloom and Gaussian blur.
Each pixel is also drawn to a temporary Vector3 buffer now instead of
directly to the bitmap, so that intermediate changes can be made to each
pixel. Also added some Vector3 operators so that vector multiplications
and divisions are more concise.